Abstracting dynamic code to be more generic
Cleaned up the views for it, made it a generic package. I implemented
the sellers portion as another example and added the `hidden` element of
`id` back so you can match up pre-existing records easily. I've also
modified the `id` on the front end side to be either the real `id` or an
ID prefixed with `tmp-` and then a random number so that they are easy
for you to differentiate in the backend.